Delete old AWS docs
The `updateDocs` commands runs a recursive `cp` operation to move the docs into
the target docs location so that it updates existing files in that location. To
do renames, it would have to delete all files in the target directory first and
then copy. Update to `updateDocs` tool coming soon.